move AI stuff into separate dir
